Metodos Finales

+++ sistema_escolar/objetos_negocio/src/main/java/interfaces/IBecasFiltradasBO.java
@@ -13,5 +13,5 @@ public interface IBecasFiltradasBO {
 
     BecasDisponiblesResponseDTO obtenerBecasFiltradas(RequisitosDTO requisitos) throws BecaInvalidaException;
 
-    Beca obtenerBecaPorCodigo(int codigo, BecasFiltradas becasFiltradas) throws BecaInvalidaException;
+    Beca obtenerBecaPorCodigo(Long codigo, BecasFiltradas becasFiltradas) throws BecaInvalidaException;
 }
